switch mixedup test methods
