+++ oasp4j-samples/oasp4j-sample-core/src/test/java/io/oasp/gastronomy/restaurant/offermanagement/batch/impl/productimport/ProductImportJobTest.java
@@ -10,6 +10,8 @@ import org.junit.Test;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecution;
+import org.springframework.batch.core.JobParameters;
+import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.test.context.ContextConfiguration;
 
 /**
@@ -21,13 +23,18 @@ import org.springframework.test.context.ContextConfiguration;
 public class ProductImportJobTest extends AbstractSpringBatchIntegrationTest {
 
   @Inject
-  private Job billExportJob;
+  private Job productImportJob;
 
   @Test
-  public void shouldExportBills() throws Exception {
+  public void shouldImportProducts() throws Exception {
 
-    JobExecution jobExecution = getJobLauncherTestUtils(this.billExportJob).launchJob();
+    JobParametersBuilder jobParameterBuilder = new JobParametersBuilder();
+    jobParameterBuilder.addString("drinks.file", "classpath:drinks.csv");
+    jobParameterBuilder.addString("meals.file", "classpath:meals.csv");
+    JobParameters jobParameters = jobParameterBuilder.toJobParameters();
 
+    JobExecution jobExecution = getJobLauncherTestUtils(this.productImportJob).launchJob(jobParameters);
     assertEquals(BatchStatus.COMPLETED, jobExecution.getStatus());
   }
+
 }
