+++ oasp4j-samples/oasp4j-sample-core/src/test/java/io/oasp/gastronomy/restaurant/salesmanagement/batch/impl/billexport/BillExportJobTest.java
@@ -10,8 +10,6 @@ import org.junit.Test;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.test.context.ContextConfiguration;
 
 /**
@@ -23,17 +21,13 @@ import org.springframework.test.context.ContextConfiguration;
 public class BillExportJobTest extends AbstractSpringBatchIntegrationTest {
 
   @Inject
-  private Job productImportJob;
+  private Job billExportJob;
 
   @Test
-  public void shouldImportProducts() throws Exception {
+  public void shouldExportBills() throws Exception {
 
-    JobParametersBuilder jobParameterBuilder = new JobParametersBuilder();
-    jobParameterBuilder.addString("drinks.file", "classpath:drinks.csv");
-    jobParameterBuilder.addString("meals.file", "classpath:meals.csv");
-    JobParameters jobParameters = jobParameterBuilder.toJobParameters();
+    JobExecution jobExecution = getJobLauncherTestUtils(this.billExportJob).launchJob();
 
-    JobExecution jobExecution = getJobLauncherTestUtils(this.productImportJob).launchJob(jobParameters);
     assertEquals(BatchStatus.COMPLETED, jobExecution.getStatus());
   }
 }
